refactor: extract follow-up renderers from AgentWorkspace
